Выполнил домашнее задание Полнотекстовый поиск - реализовал поиск по заголовкам и содержимому постов и хэштегам. Дооформил страницу ленты. Провел небольшой рефакторинг прошлых сценариев
