+++ add.php
@@ -3,14 +3,11 @@
 require_once('auth.php');
 require_once('helpers.php');
 
-$connect = mysqli_connect("localhost", "root", "root", "readme");
-if ($connect == false) {
-    die('Connection error: ' . mysqli_connect_error());
-}
-mysqli_set_charset($connect, "utf8");
+$connect = db_set_connection();
 
-$scriptname = pathinfo(__FILE__, PATHINFO_BASENAME);
-$add_url = '/' . $scriptname;
+$path = (pathinfo(__FILE__, PATHINFO_BASENAME));
+
+$add_url = '/' . $path;
 $errors = [];
 $classname = 'text';
 $content_type_id = intval($_GET['id']) ?? '';
@@ -147,6 +144,6 @@ $active_form = include_template('adding-post-forms/adding-' . $classname . '-for
 
 $adding_post = include_template('adding-post.php', ['active_form' => $active_form, 'types' => $sql_types, 'url' => $add_url, 'classname' => $classname]);
 
-$add_post_layout = include_template('layout.php', ['content' => $adding_post, 'title' => 'readme: добавление публикации', 'is_auth' => $is_auth, 'user_name' => $user_name, 'avatar' => $user_avatar]);
+$layout = include_template('layout.php', ['content' => $adding_post, 'title' => 'readme: добавление публикации', 'is_auth' => $is_auth, 'user_name' => $user_name, 'avatar' => $user_avatar, 'path' => $path]);
 
-print($add_post_layout);
+print($layout);
