+++ feed.php
@@ -4,13 +4,40 @@ require_once('auth.php');
 require_once('helpers.php');
 
 $path = (pathinfo(__FILE__, PATHINFO_BASENAME));
+$url = "/" . $path;
+$type_classname = '';
 
 $connect = db_set_connection();
-$user_name = $user['user_login'];
+if (isset($_GET['type'])) {
+    $type_classname = htmlspecialchars($_GET['type']);
+};
+
+if ($type_classname) {
+    $query_condition = "WHERE classname = '$type_classname'";
+} else {
+    $type_classname = '';
+    $query_condition = "";
+};
+
 $sql_types_query = "SELECT * FROM content_type";
 $sql_types = db_get_query('all', $connect, $sql_types_query);
 
-$feed = include_template('feed-layout.php', ['types' => $sql_types]);
-$feed_layout = include_template('layout.php', ['content' => $feed, 'title' => 'readme: моя лента', 'is_auth' => $is_auth, 'user_name' => $user_name, 'avatar' => $user_avatar, 'path' => $path]);
+$sql_posts_query = "SELECT post.*, u.user_login, u.avatar, ct.classname,
+(SELECT COUNT(id) FROM comments WHERE comments.post_id = post.id) AS total_comm,
+(SELECT COUNT(id) FROM likes WHERE likes.post_id = post.id) AS total_likes
+FROM post
+    LEFT JOIN user u ON post.user_id = u.id
+    LEFT JOIN content_type ct ON type_id = ct.id $query_condition";
+$sql_posts = db_get_query('all', $connect, $sql_posts_query);
 
-print($feed_layout);
+
+$sql_hashtag_query = "SELECT hashtag.*, hp.post_id
+                      FROM hashtag
+                            LEFT JOIN hashtags_posts hp ON hashtag.id = hp.hashtag_id";
+$sql_hashtags = db_get_query('all', $connect, $sql_hashtag_query);
+
+$feed_layout = include_template('feed-layout.php', ['types' => $sql_types, 'posts' => $sql_posts, 'hashtags' => $sql_hashtags, 'type_classname' => $type_classname, 'url' => $url, 'path' => $path]);
+
+$layout = include_template('layout.php', ['content' => $feed_layout, 'title' => 'readme: моя лента', 'is_auth' => $is_auth, 'user_name' => $user_name, 'avatar' => $user_avatar, 'path' => $path]);
+
+print($layout);
