+++ post.php
@@ -5,8 +5,9 @@ require_once('helpers.php');
 
 $connect = db_set_connection();
 
-$scriptname = pathinfo(__FILE__, PATHINFO_BASENAME);
-$post_url = '/' . $scriptname;
+$path = (pathinfo(__FILE__, PATHINFO_BASENAME));
+
+$post_url = '/' . $path;
 
 if (isset($_GET['post_id'])) {
     $post_id = intval($_GET['post_id']);
@@ -31,6 +32,11 @@ if ($post_id == $sql_post_id['id']) {
 };
 $sql_post = db_get_query('assoc', $connect, $sql_post_query);
 
+$sql_hashtag_query = "SELECT hashtag.*, hp.post_id
+                      FROM hashtag
+                            LEFT JOIN hashtags_posts hp ON hashtag.id = hp.hashtag_id";
+$sql_hashtags = db_get_query('all', $connect, $sql_hashtag_query);
+
 $current_user = $sql_post['user_id'];
 
 $sql_total_posts = get_total_from_db ('id', 'post', 'user_id', $current_user, $connect);
@@ -39,8 +45,8 @@ $sql_total_subs = get_total_from_db ('user_id', 'subscribtions', 'to_user_id', $
 
 $active_post = include_template('post-' . $sql_post['classname'] . '.php', ['post' => $sql_post]);
 
-$post_details = include_template('post-details.php', ['active_post' => $active_post, 'post' => $sql_post, 'totalpost' => $sql_total_posts, 'likes' => $sql_total_likes, 'subs' => $sql_total_subs]);
+$post_layout = include_template('post-layout.php', ['active_post' => $active_post, 'post' => $sql_post, 'totalpost' => $sql_total_posts, 'likes' => $sql_total_likes, 'subs' => $sql_total_subs, 'hashtags' => $sql_hashtags]);
 
-$post_layout = include_template('layout.php', ['content' => $post_details, 'title' => 'readme: публикация', 'is_auth' => $is_auth, 'user_name' => $user_name, 'avatar' => $user_avatar]);
+$layout = include_template('layout.php', ['content' => $post_layout, 'title' => 'readme: публикация', 'is_auth' => $is_auth, 'user_name' => $user_name, 'avatar' => $user_avatar, 'path' => $path]);
 
-print($post_layout);
+print($layout);
